ST-56: Trade list

+++ app/src/main/java/com/uguinformatica/bluemoon/androidapp/domain/models/Tradeable.kt
@@ -1,5 +1,3 @@
 package com.uguinformatica.bluemoon.androidapp.domain.models
 
-import androidx.compose.ui.text.font.FontWeight
-
-data class Tradeable(var weight: Float, var description: String, var sellPrice: Float, var sliverType: SilverType)
+data class Tradeable(var weight: Double, var description: String, var sellPrice: Double, var sliverType: SilverType)
